Connected db and Tested
Added in the Flask Form inputs instead of HTML ones. Made other necessary modifications to the HTML files to make them work as expected when they are loaded or sending data with 'POST'. Also Tested adding a User. Currently I can't view them in the PostgreSQL Table, but I'm able to log in and out successfully with no issues.