Got the Tasks to display in the Dashboard
Wow. I don't know where to start. I have a feeling that the flask syntax was always working, but my CSS conflicted somewhere and for some reason it wasn't showing. But before I found out, I changed my approach and send the data over in JSOn format. Parsed the JSOn object and let JavaScript handle the rest with innerHTML. I ended up with the same outcome, but this time the HTML was visible when I Inspected the element. I then commented-out the dashboard CSS and the Tasks began to show. So that's pretty much it. I'm going to create a new branch from here and focus on the Dashboard.